LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/navigation-timing-delivery-type.https_prefetch=true&bypass_cache=false-expected.txt


FAIL PerformanceNavigationTiming.deliveryType test, same origin prefetch. assert_equals: expected (string) "navigational-prefetch" but got (undefined) undefined
